fixed intial error

+++ src/SimpleStorage.sol
@@ -3,22 +3,22 @@
 pragma solidity >=0.8.17;
 
 contract SimpleStorage {
-    mapping(address => uint) StorageWitAdrress;
+    mapping(address => uint256) StorageWitAdrress;
     error UserNotFound();
     address[] NumUsers;
 
-    function addFavNum(uint _favNum) public {
+    function addFavNum(uint256 _favNum) public {
         StorageWitAdrress[msg.sender] = _favNum;
         NumUsers.push(msg.sender);
     }
 
-    function DisplayFavoriteNum() public view onlyOwner returns (uint) {
+    function DisplayFavoriteNum() public view onlyOwner returns (uint256) {
         return StorageWitAdrress[msg.sender];
     }
 
     modifier onlyOwner() {
         bool _checkUser = false;
-        for (uint i; i < NumUsers.length; i++) {
+        for (uint256 i; i < NumUsers.length; i++) {
             if (NumUsers[i] == msg.sender) {
                 _checkUser = true;
             }
